requirements.txt
asgiref==3.5.0
autopep8==1.6.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.4
Django==4.0.1
django-cors-headers==3.13.0
django-countries==7.5.1
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.6.0
MarkupPy==1.14
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.3.0
platformdirs==2.5.1
pycodestyle==2.8.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
scipy==1.9.1
six==1.16.0
sqlparse==0.4.2
tablib==3.5.0
toml==0.10.2
typing_extensions==4.1.1
tzdata==2023.3
virtualenv==20.13.4
xlrd==2.0.1
xlwt==1.3.0
